Revert to before editing the navbar
