+++ logicle/app/admin/layout.tsx
@@ -1,6 +1,6 @@
 'use client'
 import SettingsLayout from '@/app/layouts/SettingsLayout'
-import { NavEntry } from '@/components/ui/settings-navbar'
+import { NavEntry } from '@/components/ui/navbar'
 import { useTranslation } from 'react-i18next'
 import { Environment, useEnvironment } from '../context/environmentProvider'
